Refactor main.go into cmd, utils packages and the github client.
